requirements.txt
tweepy
python-dotenv
Faker
